Fix and clean up collision tests
